+++ site/src/pages/apis/index.tsx
@@ -4,31 +4,89 @@ import Layout from '@theme/Layout';
 import Link from '@docusaurus/Link';
 import useDocusaurusContext from '@docusaurus/useDocusaurusContext';
 import styles from '../index.module.css';
-import ApiPages from '../../components/apiPages';
+import apiStyles from '../../css/apis.module.css';
+import ItemGrid from '../../components/gridItem'
 
-function HomepageHeader() {
-  const {siteConfig} = useDocusaurusContext();
-  return (
-    <header className={clsx('hero hero--primary', styles.heroBanner)}>
-      <div className="container">
-        <h1 className="hero__title">Bandwidth APIs</h1>
-      <p className="hero__subtitle">Build with Bandwidth</p>
-      </div>
-    </header>
-  );
+const ApiList = [
+    {
+        title: 'Numbers',
+        Svg: require('../../../static/img/icon-numbers-complex.svg').default,
+        style: apiStyles.apiSvg,
+        link: '/apis/numbers'
+    },
+    {
+        title: 'Phone Number Lookup',
+        Svg: require('../../../static/img/icon-numbers.svg').default,
+        style: apiStyles.apiSvg,
+        link: '/apis/number-lookup'
+    },
+    {
+        title: 'Voice',
+        Svg: require('../../../static/img/icon-voice-complex.svg').default,
+        style: apiStyles.apiSvg,
+        link: '/apis/voice'
+    },
+    {
+        title: 'Messaging',
+        Svg: require('../../../static/img/icon-messaging-complex.svg').default,
+        style: apiStyles.apiSvg,
+        link: '/apis/messaging'
+    },
+    {
+        title: 'International Messaging',
+        Svg: require('../../../static/img/icon-messaging.svg').default,
+        style: apiStyles.apiSvg,
+        link: '/apis/messaging-international'
+    },
+    {
+        title: 'Multi-Factor Authentication',
+        Svg: require('../../../static/img/icon-mfa.svg').default,
+        style: apiStyles.apiSvg,
+        link: '/apis/multifactorauth'
+    },
+    {
+        title: 'WebRTC',
+        Svg: require('../../../static/img/icon-WebRTC.svg').default,
+        style: apiStyles.apiSvg,
+        link: '/apis/webrtc'
+    },
+    {
+        title: 'DASH',
+        Svg: require('../../../static/img/icon-emergency-dashboard.svg').default,
+        style: apiStyles.apiSvg,
+        link: '/apis/dash'
+    },
+    {
+        title: 'DASH Notifications',
+        Svg: require('../../../static/img/icon-emergency-calling.svg').default,
+        style: apiStyles.apiSvg,
+        link: '/apis/dash-notifications'
+    },
+];
+
+function Header() {
+    const { siteConfig } = useDocusaurusContext();
+    return (
+        <header className={clsx('hero hero--primary', styles.heroBanner)}>
+            <div className="container">
+                <h1 className="hero__title">Bandwidth APIs</h1>
+                <p className="hero__subtitle">Build with Bandwidth</p>
+            </div>
+        </header>
+    );
 }
 
-export default function Home() {
-  const {siteConfig} = useDocusaurusContext();
-  return (
-    <Layout
-      title="Bandwidth APIs"
-      description="Bandwidth API Reference, documentation, SDKs, guides, examples and more. Get everything you need to build with Bandwidth."
-      keywords="Bandwidth,API">
-      <HomepageHeader />
-      <main>
-        <ApiPages />
-      </main>
-    </Layout>
-  );
+export default function Apis() {
+    const { siteConfig } = useDocusaurusContext();
+    return (
+        <Layout
+            title="Bandwidth APIs"
+            description="Bandwidth API Reference, documentation, SDKs, guides, examples and more. Get everything you need to build with Bandwidth."
+            keywords="Bandwidth,API">
+            <Header />
+            <main>
+                <ItemGrid itemList={ApiList} className={apiStyles.apis}/>
+            </main>
+        </Layout>
+    );
 }
